+++ ocrd_models/ocrd_models/ocrd_file.py
@@ -1,7 +1,9 @@
 """
 API to ``mets:file``
 """
-import os
+from os.path import splitext, basename
+
+from ocrd_utils import is_local_filename, get_local_filename
 
 from .ocrd_xml_base import ET
 from .constants import NAMESPACES as NS, TAG_METS_FLOCAT, TAG_METS_FILE
@@ -15,7 +17,7 @@ class OcrdFile():
     #  def create(mimetype, ID, url, local_filename):
     #      el_fileGrp.SubElement('file')
 
-    def __init__(self, el, mimetype=None, instance=None, local_filename=None, mets=None):
+    def __init__(self, el, mimetype=None, instance=None, local_filename=None, mets=None, url=None):
         """
         Args:
             el (LxmlElement):
@@ -32,15 +34,12 @@ class OcrdFile():
         self._instance = instance
         self.mets = mets
 
-        if self.url:
-            if not self.url.startswith('http://') and \
-                not self.url.startswith('https://'):
-                self.local_filename = self.url
-            if self.url.startswith('file://'):
-                self.local_filename = self.url[len('file://'):]
+        if url:
+            self.url = url
 
-        #  if baseurl and not local_filename and '://' not in self.url:
-        #      self.local_filename = '%s/%s' % (baseurl, self.url)
+        if not(local_filename):
+            if self.url and is_local_filename(self.url):
+                self.local_filename = get_local_filename(self.url)
 
     def __str__(self):
         """
@@ -62,12 +61,12 @@ class OcrdFile():
         """
         Get the ``os.path.basename`` of the local file, if any.
         """
-        return os.path.basename(self.local_filename)
+        return basename(self.local_filename if self.local_filename else self.url)
 
     @property
     def extension(self):
-        basename, ext = os.path.splitext(self.basename)
-        if basename.endswith('.tar'):
+        _basename, ext = splitext(self.basename)
+        if _basename.endswith('.tar'):
             ext = ".tar" + ext
         return ext
 
@@ -139,7 +138,10 @@ class OcrdFile():
         """
         The ``USE`` attribute of the containing ``mets:fileGrp``
         """
-        return self._el.getparent().get('USE')
+        parent = self._el.getparent()
+        if parent is not None:
+            return self._el.getparent().get('USE')
+        return 'TEMP'
 
     @property
     def url(self):