+++ ocrd_utils/ocrd_utils/constants.py
@@ -7,6 +7,7 @@ __all__ = [
     'VERSION',
     'MIMETYPE_PAGE',
     'EXT_TO_MIME',
+    'MIME_TO_EXT'
 ]
 
 VERSION = get_distribution('ocrd_utils').version
@@ -20,4 +21,13 @@ EXT_TO_MIME = {
     '.jpg': 'image/jpg',
     '.jpeg': 'image/jpg',
     '.xml': MIMETYPE_PAGE
+}
+
+MIME_TO_EXT = {
+    'image/tiff': '.tif',
+    'image/png': '.png',
+    'image/jpg': '.jpg',
+    'image/jpeg': '.jpg',
+    MIMETYPE_PAGE: '.xml',
+    'application/alto+xml': '.xml',
 }